+++ materials/carcass/categories/delete.php
@@ -1,11 +1,13 @@
 <?php
-    include '../../../db.php';
-    /** @var mysqli $conn */
-    $id = (int) ($_GET['id'] ?? 0);
-    mysqli_query(
-        $conn,
-        "DELETE FROM carcass_categories
-        WHERE id = '$id'"
-    );
-    header('Location: manage.php');
-?>
+include '../../../db.php';
+if(!can('carcass_delete')){
+    die('Access Denied');
+}
+/** @var mysqli $conn */
+$id = (int) ($_GET['id'] ?? 0);
+mysqli_query(
+    $conn,
+    "DELETE FROM carcass_categories
+    WHERE id = '$id'"
+);
+header('Location: manage.php');
