+++ materials/carcass/categories/manage.php
@@ -2,6 +2,9 @@
 <?php
 include '../../../includes/auth.php';
 include '../../../db.php';
+if(!can('carcass_view')){
+    die('Access Denied');
+}
 /** @var mysqli $conn */
 include '../../../includes/header.php';
 include '../../../includes/sidebar.php';
@@ -15,7 +18,9 @@ $query = mysqli_query(
 <div class="page-card">
     <div class="page-header">
         <h1>Carcass Categories</h1>
-        <a href="create.php" class="theme-btn">+ Add Category</a>
+        <?php if(can('carcass_create')){ ?>
+            <a href="create.php" class="theme-btn">+ Add Material</a>
+        <?php } ?>
     </div>
     <table class="custom-table">
         <thead>
@@ -39,8 +44,12 @@ $query = mysqli_query(
                         <?php } ?>
                     </td>
                     <td>
-                        <a href="edit.php?id=<?= $row['id'] ?>" class="edit-btn">Edit</a>
-                        <a href="delete.php?id=<?= $row['id'] ?>" class="delete-btn">Delete</a>
+                        <?php if(can('carcass_edit')){ ?>
+                            <a href="edit.php?id=<?= $row['id'] ?>" class="edit-btn">Edit</a>
+                        <?php } ?>
+                        <?php if(can('carcass_delete')){ ?>
+                            <a href="delete.php?id=<?= $row['id'] ?>" class="delete-btn" onclick="return confirm('Delete Category?')">Delete</a>
+                        <?php } ?>
                     </td>
                 </tr>
             <?php } ?>
